StudentConnect-main/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cloudinary==1.36.0
python-decouple==3.8
pydantic==2.3.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
email-validator==2.1.0